Handle "(revision )" for source files.
Our svn revision parsing code already had a special case for the pretty wacky
"(revision )", but it only worked when this was seen for the new revision. If
we got this for the source revision, we'd hit various errors. I've changed the
logic to report this as PRE_CREATION, which shouldn't break the old IntelliJ
diffs but should work for whatever wonky tool was used for bug 2632.

Testing done: Ran unit tests.