Fix "From" field in emails for --submit-as
When sending emails for review requests, reviews, and replies, we'd send it
from the request user, rather than the owner of the relevant object. This made
using --submit-as pretty ugly.

At some point in the future, we want to enable many users to collaborate on the
same review request, at which point we'll have to revisit this, but for now,
fix the bugs.

Testing done: Ran unit tests

Fixes bugs 1795 and 2500.
Reviewed at https://reviews.reviewboard.org/r/4418/