Improve i18n a bit.
I'm working on integrating the Italian localization that we recieved at
https://reviews.reviewboard.org/r/4298/

There are a few things that were getting in the way here. We have some old i18n
scripts in contrib/ which mostly work, but should be removed in favor of the
django-admin.py versions. The django-admin.py version of makemessages was
erroring out in a couple places where we were using positional format strings
with multiple arguments, instead of parameter names. I've fixed these up.

There are also several strings which aren't marked for translation. A lot of
these are likely coming from javascript, but one that was immediately obvious
was the dashboard's sidebar.

Testing done:
Integrated the contributed Italian localization files, and ran
compilemessages. Set my browser to prefer Italian, and saw that most of
the UI was now translated. Neat.

Reviewed at https://reviews.reviewboard.org/r/4426/