Make use of Beanstalk's new blob API for fetching files.
Beanstalk kindly added a blob API for us so that we could easily fetch
the contents of files without need the base_commit_id. That gives our
users the choice to continue using older versions of RBTools or to
upload a standard 'git diff' if they choose.

We now use the new blob API when fetching the raw contents of a file, or
when we need to check file existence without having a base_commit_id.

Unit tests were updated to handle the new flows, and to not expect
errors in the case of Git + no base commit ID.