Add en, it_IT, and zh_TW locale files.
This change pulls in the Italian and Chinese localizations that have been
contributed. Some of the translated strings are out of date (and thus marked
"fuzzy"), and there are still a number of strings, mostly from JavaScript, that
aren't translatable, but this is a pretty good start.

Testing done:
Set my browser language preferences to prefer Italian and (Traditional) Chinese
over English. Reloaded, and saw my dashboard and review requests in other
languages.