Mark strings in javascript for translation.
Django provides a nice, gettext-like API for localizing strings in javascript
files. Now that we've integrated the jsi18n view, we can mark these strings so
that people can translate them.

The only unfortunate thing about this is that there's no good way to wrap
strings and have them concatenate (like there is with python or C), so for
longer strings that we want to localize, they have to be all together on one
line.

Testing done:
- Extracted messages and saw that djangojs.po had all the expected strings.
- Added some fake italian and chinese versions of the "An issue was opened"
  string and saw that when my browser was set to prefer those languages, I saw
  the translated string.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/4453/