Fix S3 querystring auth.
The boto-based S3 backed defaults to using query-string authentication, which
causes the URL to expire after a while. This is meant to prevent hot-linking,
but in our case, it's easy to have it not quite sync up with our caching,
ending up with a cached link which is then expired. This change turns off
querystring auth in the storage backend.

Testing done:
Configured my devserver to use S3, and uploaded a file. Checked and saw that
the relevant URLs generated no longer contained the various querystring
parameters.

Reviewed at https://reviews.reviewboard.org/r/4457/