Handle ImproperlyConfigured when trying to load the S3 backend.
If django-storages is installed, but boto is not, it will raise an
ImproperlyConfigured exception when trying to do the import, which we didn't
catch. This caused the admin form to error out.

Testing done:
Removed the boto package from my site-packages and tried to load the file
storage settings page. When I selected "Amazon S3", saw the error message from
the exception.

Fixes bug 3062.

Reviewed at https://reviews.reviewboard.org/r/4463/