Add review UIs to the diff viewer.
We now display review UIs in the diff viewer for file attachments that
have an inline-capable review UI. This takes the place of the thumbnail
that we would otherwise render.

When displaying an inline-capable review UI, the "Review" button goes
away, and the review UI is just rendered inline. It can be commented on,
just like normal.

At the moment, both the original and modified versions each have a
review UI capable of commenting. It's unclear whether the left-hand side
should allow commenting on a review UI, but right now, it's there. Extra
work would need to be done to tell the AbstractReviewables that they're
not allowed to add commenting capabilities.

The next step is to add diffable file attachment support to review UIs.