Add icons for search, jump-to, start to use them in the new review request page.
This change adds new icons to our spritesheet for searching the repository
list, and a sort of jump-to icon to denote revisions which already have an
existing review request. The jump-to icon is used as part of this change.
Search will be used in a later change.

Testing done:
- Created a few review requests for committed revisions, then went back to the
  new review request page. Saw the jump-to icon in the right spot.
- Added a div for the search icon using the chrome devtools and saw that it
  looked nice in context.

Reviewed at http://reviews.reviewboard.org/r/4470/