If there's only one repository, auto-select it for new Review Requests.
A lot of our users have only a single repository. In the new review request
page, if we're in this situation, auto-select the repository, to reduce one
click.

Testing done:
Set visible=False for all but one of my repositories. Went to /r/new and saw it
automatically select the remaining one. Un-hid my repositories and saw the
"Select a repository" message.

Reviewed at https://reviews.reviewboard.org/r/4472/