Fix alignment of the file attachment thumbnail overlay.
We put a clickable overlay on top of file attachment thumbnails, to ensure that
clicking anywhere within the area of the attachment will redirect to the review
UI (instead of selecting text or accidentally activating any links that might
be part of the thumbnail, e.g. for text file types). This overlay wasn't set to
have a left alignment of 0, so its left edge was aligned with the thumbnail
itself. This worked for some types (like text) but failed when we had
thumbnails that didn't stretch the full width of the box.

Testing done:
Verified alignment in Chrome and Firefox.

Reviewed at https://reviews.reviewboard.org/r/4475/