Remove preloading of the first file in the diff viewer.
The diff viewer preloaded the first file in an attempt to reduce
perceived load time by ensuring that upon the first load, there was
something to review.

This is great in theory, but it doesn't really survive some of the new
work. When we reach the point of being able to offload diff generation,
we won't be able to ensure the first file is loaded, creating an
inconsistent experience when enabling such a feature.

Also, some of the new review UI code gets much more complex with
preloading, since we need to handle some logic in two different places.

So, this rips that out. We'll need to get a feel for the loading time
perception of this, and see if there are any decent workarounds.