Build and ship compiled locale files.
This adds another compilation step to the egg build: locale files. We keep the
.po files checked in to git, but we don't want to check in .mo files because
they're big binary blobs that may change frequently, and git doesn't do so well
with that.

This requires that msgfmt from the gettext package is installed somewhere in
the path. On OSX, I've installed it via homebrew, but that doesn't add it to
the user path by default because of some conflicts.

Testing done:
- Built a .egg and installed it. Verified that both .po and .mo files were
  included in the correct location.
- Changed my PATH so that msgfmt could not be found. Saw the bdist_egg command
  error out.