Split RB.linkifyText into separate pieces.
The linkifyText method does a bunch of things. It starts out by escaping out <
and > characters (which doesn't have much to do with linkifying), and then
proceeds to create links for any full URLs, /r/### review requests, and bug
numbers.

When we integrate markdown support for text boxes, we don't want the <>
escaping, and full URLs will be handled by the markdown parser. We therefore
want to be able to call subsets of this functionality in addition to the whole.

Testing done: Ran js-tests
Reviewed at https://reviews.reviewboard.org/r/4489/