Make field formatters in ReviewRequestEditorView set the element's html.
ReviewRequestEditorView's field definitions have an abstract "formatter"
concept, which was a function that would mutate the value before setting it in
the DOM. This change moves the responsibility for calling jQuery.html() into
the formatter functions themselves. The motivation for this is that the coming
markdown changes will need to do some fancier stuff including accessing data
attributes on the element (to check if it should be formatted as rich text),
and setting the rich-text class once it's been formatted.

Testing done: Ran js-tests.
Reviewed at https://reviews.reviewboard.org/r/4490/