Make linkifyReviewRequests and linkifyBugs produce markdown output.
This change adds a 'markdown' parameter to the review requests and bugs linkify
functions that enables them to produce output in markdown format (as opposed to
HTML).

While I was doing this, I noticed that writing "issue #2, issue #4" would
include the comma in the first bug number link. I've fixed the regex to handle
this case properly (commas are pretty unlikely to be part of a bug number), and
added a relevant test.

Testing done:
- Used this in conjuction with the markdown changes.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/4491/