Make some CSS selectors more specific to avoid style bleed into descendents.
The list of replies to a comment on the review request page are arranged
semantically, using ul>li. Unfortunately, the reply-comments style bled through
to *all* li descendants of the top-level list, and not just immediate children.
This meant that when markdown formatting used a list (for actual lists, or the
way that it formats code samples), it would have all the junk used to display
the replies.

Similarly, a couple of the elements inside the change description banners at
the top of the page had loose selectors which caused styles to bleed into the
<pre> unnecessarily.

Testing done:
Saw that lists inside the .reviewtext and change description elements didn't
inherit a bunch of junky styles.

Reviewed at https://reviews.reviewboard.org/r/4492/