Fix the field name for changedescription in drafts.
The DraftReviewRequest model was trying to pull out "change_description" from
the JSON payload for the draft's change description, but the API field name is
'changedescription'. We didn't notice this because it wasn't actually used
anywhere (the change description was saved but never re-read).

Testing done:
- Ran js-tests
- Used this in conjunction with a change that adds a formatter to the
  changedescription field in the review request editor, causing the new value
  to be read when we save the draft.

Reviewed at https://reviews.reviewboard.org/r/4496/