Don't auto-open the change description editor when loading a draft.
One thing that's always annoyed me about loading a public review request with a
draft is that the change description editor box auto-opens. Until now this has
been a mere annoyance, but with the markdown change, it's downright ugly. This
change kills that behavior.

Testing done:
Loaded a review request with a draft and saw the change description stay where
it was when the page loaded.

Reviewed at https://reviews.reviewboard.org/r/4497/