Move i18n building into a script to fix deployment issues.
For some reason, doing a 'fab deploy' results in the django import in
Build18N to fail. The code doesn't seem to be running with the
PYTHONPATH we'd expect.

This change moves the i18n building logic out into a script, just like
media building. The script is then able to run in the context of the
virtualenv, and not in whatever screwed up environment the class is
otherwise running in.