Don't set an index on FileAttachment.repo_path.
MySQL has a rather low-ish limit for indexes on Unicode tables. We hit
that with the index for the new repo_path field.

This fixes the model and the initial evolution to not set an index. A
new evolution was added to fix the databases synced since the original
change, so that our state matches up, and otherwise will do nothing to
an existing database.