Call success and update draft field when setting the close description.
The changeDescription field works kind of strangely when a review request is
closed. Rather than setting a field in a draft and asking it to save, it calls
close() with the new text. Unfortunately, the changeDescription field in
reviewRequest.draft was getting out of sync (since it wasn't updated), and the
success callback was never triggered. This fixes both those issues.

Testing done: Used in conjunction with my markdown change.
Reviewed at https://reviews.reviewboard.org/r/4504/