Fix ReviewBoxListView iteration over changedescs.
Our markup for change description boxes is a little weird, where we end up with
multiple nested elements with the 'changedesc' class. ReviewBoxListView was
iterating over everything with that class to set up the collapsable boxes,
which worked okay, but as soon as I tried to add something else to it, things
got weird.

This changes ReviewBoxListView.el to point to #content, and then iterates only
over the children.

Testing done: Used in conjunction with the markdown change.
Reviewed at https://reviews.reviewboard.org/r/4505/