Store the docs database as fixtures instead of an sqlite database.
The docs database was painful to merge and update, and sometimes stray
items (from operations that deal with in-database caches) would get
committed. It also outright failed when there were evolutions that
needed to be applied, always.

Now, building any docs target for the manual will rebuild the docs db
from a fixture, if the fixture has changed or if the database is
missing. When adding new items to the database, a modified docs.db can
be dumped using ./scripts/dumpdata.sh.

This should help going forward with some annoyances with our docs
builds.