Remove the trailing blank line in markdown text.
The markup() function returns a string with an extra blank line (as
every HTML tag is outputted with a trailing newline). We don't really
want this, since it leaves a blank line in each <pre>, so check if
there's a newline and strip it away. This brings the look to how we had
things pre-markdown.