Improve the output of the API docs.
The API docs need a lot of work, and this is one step in that direction.

We now longer show sample XML payloads. While they work, it's preferable
that people just use JSON, so advertise that instead. This has the
benefit of lengthening some very large API doc pages.

The examples now have a sample curl command line for getting the
resource. It also shows the headers that come back.

The description of the resource is now at the top, before the other
details (like URL, child resources, etc.)

Some bullet points have been removed where they just added clutter.

There are some fixes for table layouts. Mostly, this is dealing with
adding CSS classes, which will work with stylesheet changes for the RB
website.