Add a new base TestCase class with helpful factory methods.
This introduces a new reviewboard.testing app with a new base TestCase
class that's intended to be used in other tests. This class provides a
handful of methods for creating various models we use in tests, along
with default data (which can be overridden by the callers).

This will allow us to reduce our dependence on most fixtures, which will
speed up tests, and make tests more readable.