Move away from most fixtures on all non-webapi apps.
This removes the usage of the test_reviewrequests and most of the
test_scmtools fixtures from the non-webapi unit tests. Instead, these
test suites now use the new TestCase class's utility methods for
creating test objects.

With the test suite as it originally was, on my machine, it took roughly
159 seconds to complete. With this change, it's down to ~143. After my
upcoming change, which also removes data from the test_scmtools fixture,
the time drops to ~88.