Add team support for Bitbucket repositories.
The new Bitbucket hosting service rewrite, which brought along Git
support, only worked for personal acounts. It didn't support team
accounts.

Now, Bitbucket supports two types of plans: Personal, and Team.
This works much like the various plans for GitHub. Depending on the one
selected, we configure requests against the repositories differently.
For personal accounts, the repository owner name in the URL is the
username of the owner, For team accounts, it's the team name, which is
requeste during configuration.

This works with existing configurations. It's assumed that any Bitbucket
repository configuration without an explicit plan set is a personal
account.