Update for PUT payload changes in the test client.
In 1.5, django.test.Client changed the behavior of PUT, OPTIONS, and DELETE. In
1.4.x and prior, OPTIONS and DELETE behaved like GET (encoded any data in the
query string), and PUT behaved like POST (encoding data as multipart). These
are now very generic, and it's up to us to manually encode the data and set the
content-type for PUT requests if we want it to get to the server.

Testing done: Ran unit tests.
Reviewed at https://reviews.reviewboard.org/r/4620/