Raise an error when uploading a non-git diff to a git repository.
When uploading a non-git diff (whether an image file or anything else),
GitDiffParser would assume it's just an empty file. We now sanity check
this better by seeing if we parsed any files, and if not, if we're
dealing with a non-empty file. If no files were parsed and the file is
not just empty or whitespace, we assume it's some unknown file type.