Fix the subject in test_put_reviewrequestdraft_publish
In Django 1.5.x, the locmem email backend will trigger header validation and
generate a message ID. This means that the subject when publishing a second
draft will be "Re: ...". Our previous implementation of the assertion was
actually incorrect, but it took a while to convince myself of that.

Testing done: Ran unit tests.
Reviewed at http://reviews.reviewboard.org/r/4621/