Return better output on webapi doc parse_text failures.
When docs are missing from a class or function, parse_text will fail
accessing a function on a None. That's not useful, and the error output
doesn't tell us what happened or where.

Now, most callers provide a 'where' string that's used for any assertion
messages to better let us know what was missing.