Don't show "Show changes" when there is only one diff revision
Added logic to review_detail.html so that if the past revision number of a diff
is 0 (when the first revision of the diff has been added after a review request
has been published) so that the "Show changes" link would not be displayed.

Testing done:
Reproduced the bug by publishing a review request, then adding a diff and
verifying that the "Show changes" link does not appear. Added a second diff to
the same review request and verified that the "Show changes" link does appear
as it should for the second diff.

Reviewed at https://reviews.reviewboard.org/r/4613/
Fixes bug 2901