Fix issues with docs generation on Django 1.5.
Docs generation on Django 1.5 didn't work due to the 'reset' management
command being removed. We used this command to remove the SCMTool data
that gets installed by default on a database, but another option is to
just run with RUNNING_TEST=1, which will bypass that.

The payload data also needed to be split on \r\n\r\n instead of \n\n.

Along with this, to help with error output, I added another where=
for a parse_text call, and made the assertion check less strict.