Fix docs generation properly.
My previous fix for docs generation didn't actually work. It just
appeared to, due to my test conditions.

We now set RUNNING_TEST = True in settings, which is where it should
go. The Makefile no longer has to set it.

make clean also now removes the database, and this is called when
fixtures change.